refactor(ipmb): pr #6 - minor reactorings
